+++ changes.md
@@ -1,7 +1,9 @@
 # Log of major changes
 v0.2.0
 List major changes
-
+* HypyDDCore: Change dep_corr to 0.0 km as the default for HypoDD.
+* workflow: Added dep_corr variable in workflow as an input parameter.
+* example/run_HypoXPy_workflow.ipynb: added dep_corr variable as an input.
 
 -----------
 v0.1.0